Mode switching protections
In the design of the mode switching backend, I have forseen the possibility of the end user pressing a button to switch modes while still holding down some of the buttons involved in the previous mode. I made some enhancements so now if this happens then any paradigm associated with those buttons will remain active until all the buttons of assocatied with that paradigm have been released. These paradigms are considered to be in a "pending" state. This provision avoids keys being virtually stuck down as a result of a paradigm being deactivated in the middle of being realized. If a new mode is activated while there are paradigms still pending deactivation, then if there are any paradigms of the new mode that conflict with the pending paradigms, these paradigms will enter a "waiting" state until the conflicting pending paradigms finally deactivate. However note that if a mode, M, is activated, with paradigm, P, modifying a button, B, which the end user is holding down, and B is not part of any pending paradigm (i.e. it was not a modified button prior to activation of M), then P will immediately go in to affect (i.e. it will not wait for the user to release B). This could have unexpected consequences, such as an unintended call P.LP_longUp(). Furture versions may remedy this by checking the down state of buttons prior to paradigm activation.